Fix template dag not parsed.
The airflow webserver only parses files that it determines have dags defined by looking for the airflow.DAG method.  If a file does not import this method, it does not get parsed in the initialairflow-webserver setup, and even if dags are being defined behind a layer of abstraction, they will not show up on the airflow dashboard.